Create worksheet-selectionchange
This code was written with an actual County/State tag in place of [COUNTY]/[STATE]. I edited this information out just to be safe. This code is entirely my own, and was in use by only myself and the team members I managed. At no point did this code belong to the county I worked for, nor were there any issues with me taking the code home after my contract expired. I am not sure if this is still in use or not, as I can no longer maintain the code for whatever new projects they may have shifted to.